requirements.txt
notion-client
SpeechRecognition
PyAudio
pyttsx3
spacy
python-dotenv
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
dateparser